可以自动玩了,但是还没盈利

+++ src/core/utils.py
@@ -25,4 +25,4 @@ def get_randomized_amount(amount, jitter=0.05):
     """
     Adds a small random jitter to an amount (e.g., 3BB becomes 2.9BB - 3.1BB).
     """
-    return amount * (1 + random.uniform(-jitter, jitter))
+    return int(amount * (1 + random.uniform(-jitter, jitter)))
